+++ src/main.rs
@@ -25,6 +25,11 @@ extern crate termios;
 extern crate winapi;
 extern crate cpp_demangle;
 extern crate rand;
+extern crate serde;
+#[macro_use]
+extern crate serde_derive;
+extern crate serde_json;
+
 extern crate remoteprocess;
 
 mod config;
@@ -39,8 +44,9 @@ mod python_spy;
 mod stack_trace;
 mod console_viewer;
 mod flamegraph;
-mod utils;
+mod speedscope;
 mod timer;
+mod utils;
 mod version;
 
 use std::io::Read;
@@ -53,6 +59,7 @@ use failure::Error;
 use python_spy::PythonSpy;
 use stack_trace::StackTrace;
 use console_viewer::ConsoleViewer;
+use config::{Config, FileFormat};
 
 fn print_traces(traces: &[StackTrace], show_idle: bool) {
     for trace in traces {
@@ -95,7 +102,7 @@ fn permission_denied(err: &Error) -> bool {
 
 fn sample_console(process: &mut PythonSpy,
                   display: &str,
-                  config: &config::Config) -> Result<(), Error> {
+                  config: &Config) -> Result<(), Error> {
     let rate = config.sampling_rate;
     let mut console = ConsoleViewer::new(config.show_line_numbers, display,
                                          &format!("{}", process.version),
@@ -124,11 +131,59 @@ fn sample_console(process: &mut PythonSpy,
     Ok(())
 }
 
+pub trait Recorder {
+    fn increment(&mut self, traces: &[StackTrace]) -> Result<(), Error>;
+    fn write(&self, w: &mut std::fs::File) -> Result<(), Error>;
+}
 
-fn sample_flame(process: &mut PythonSpy, filename: &str, config: &config::Config) -> Result<(), Error> {
+impl Recorder for speedscope::Stats {
+    fn increment(&mut self, traces: &[StackTrace]) -> Result<(), Error> {
+        for trace in traces {
+            self.record(trace)?;
+        }
+        Ok(())
+    }
+    fn write(&self, w: &mut std::fs::File) -> Result<(), Error> {
+        self.write(w)
+    }
+}
+
+impl Recorder for flamegraph::Flamegraph {
+    fn increment(&mut self, traces: &[StackTrace]) -> Result<(), Error> {
+        Ok(self.increment(traces)?)
+    }
+    fn write(&self, w: &mut std::fs::File) -> Result<(), Error> {
+        self.write(w)
+    }
+}
+
+pub struct RawFlamegraph(flamegraph::Flamegraph);
+
+impl Recorder for RawFlamegraph {
+    fn increment(&mut self, traces: &[StackTrace]) -> Result<(), Error> {
+        Ok(self.0.increment(traces)?)
+    }
+
+    fn write(&self, w: &mut std::fs::File) -> Result<(), Error> {
+        self.0.write_raw(w)
+    }
+}
+
+fn record_samples(process: &mut PythonSpy, config: &Config) -> Result<(), Error> {
     let max_samples = config.duration * config.sampling_rate;
 
-    let mut flame = flamegraph::Flamegraph::new(config.show_line_numbers);
+    let mut output: Box<dyn Recorder> = match config.format {
+        Some(FileFormat::flamegraph) => Box::new(flamegraph::Flamegraph::new(config.show_line_numbers)),
+        Some(FileFormat::speedscope) =>  Box::new(speedscope::Stats::new()),
+        Some(FileFormat::raw) => Box::new(RawFlamegraph(flamegraph::Flamegraph::new(config.show_line_numbers))),
+        None => return Err(format_err!("A file format is required to record samples"))
+    };
+
+    let filename = match config.filename.as_ref() {
+        Some(filename) => filename,
+        None => return Err(format_err!("A filename is required to record samples"))
+    };
+
     use indicatif::ProgressBar;
     let progress = ProgressBar::new(max_samples);
 
@@ -166,7 +221,7 @@ fn sample_flame(process: &mut PythonSpy, filename: &str, config: &config::Config
 
         match process.get_stack_traces() {
             Ok(traces) => {
-                flame.increment(&traces)?;
+                output.increment(&traces)?;
                 samples += 1;
                 if samples >= max_samples {
                     break;
@@ -189,9 +244,11 @@ fn sample_flame(process: &mut PythonSpy, filename: &str, config: &config::Config
         println!("{}", exit_message);
     }
 
-    let out_file = std::fs::File::create(filename)?;
-    flame.write(out_file)?;
+    {
+    let mut out_file = std::fs::File::create(filename)?;
+    output.write(&mut out_file)?;
     println!("Wrote flame graph '{}'. Samples: {} Errors: {}", filename, samples, errors);
+    }
 
     // open generated flame graph in the browser on OSX (theory being that on linux
     // you might be SSH'ed into a server somewhere and this isn't desired, but on
@@ -202,8 +259,31 @@ fn sample_flame(process: &mut PythonSpy, filename: &str, config: &config::Config
     Ok(())
 }
 
+fn run_spy_command(process: &mut PythonSpy, config: &config::Config) -> Result<(), Error> {
+    match config.command.as_ref() {
+        "dump" =>  {
+            println!("{}\nPython version {}", process.process.exe()?, process.version);
+            print_traces(&process.get_stack_traces()?, true);
+        },
+        "record" => {
+            record_samples(process, config)?;
+        },
+        "top" => {
+            let display = match config.python_program.as_ref() {
+                Some(subprocess) => subprocess.join(" "),
+                None => format!("pid: {}", config.pid.unwrap())
+            };
+            sample_console(process, &display, config)?;
+        }
+        _ => {
+            // TODO: ?
+        }
+    }
+    Ok(())
+}
+
 fn pyspy_main() -> Result<(), Error> {
-    let config = config::Config::from_commandline()?;
+    let config = config::Config::from_commandline();
 
     #[cfg(target_os="macos")]
     {
@@ -216,14 +296,7 @@ fn pyspy_main() -> Result<(), Error> {
 
     if let Some(pid) = config.pid {
         let mut process = PythonSpy::retry_new(pid, &config, 3)?;
-        if config.dump {
-            println!("{}\nPython version {}", process.process.exe()?, process.version);
-            print_traces(&process.get_stack_traces()?, true);
-        } else if let Some(ref flame_file) = config.flame_file_name {
-            sample_flame(&mut process, &flame_file, &config)?;
-        } else {
-            sample_console(&mut process, &format!("pid: {}", pid), &config)?;
-        }
+        run_spy_command(&mut process, &config)?;
     }
 
     else if let Some(ref subprocess) = config.python_program {
@@ -257,11 +330,7 @@ fn pyspy_main() -> Result<(), Error> {
         }
         let result = match PythonSpy::retry_new(command.id() as remoteprocess::Pid, &config, 8) {
             Ok(mut process) => {
-                if let Some(ref flame_file) = config.flame_file_name {
-                    sample_flame(&mut process, &flame_file, &config)
-                } else {
-                    sample_console(&mut process, &subprocess.join(" "), &config)
-                }
+                run_spy_command(&mut process, &config)
             },
             Err(e) => Err(e)
         };